Big fixed in Auto_Threshold.java
double[] tHisto = new double[iHisto.length] ; // Instead of double[] tHisto = iHisto ;
...
System.arraycopy(tHisto, 0, iHisto, 0, iHisto.length) ; //Instead of iHisto = tHisto ;

The array tHisto was in fact exactly the same as iHisto, because were both a pointer on the same array (because of `double[] tHisto = iHisto`). So I create a separate array, and I make a copy array to array.